+++ csatar-plugins/knowledgerepository/lang/en/lang.php
@@ -12,7 +12,16 @@
                     'workPlans' => 'Work Plans',
                     'methodologies' => 'Methodologies',
                 ],
-                'knowledgeRepositoryParameters' => 'Knowledge Repository Parameters',
+                'knowledgeRepositoryParameters' => [
+                    'knowledgeRepositoryParameters' => 'Knowledge Repository Parameters',
+                    'gameDevelopmentGoals' => 'Game Development Goals',
+                    'accidentRiskLevels' => 'Accident Risk Levels',
+                ],
+            ],
+            'general' => [
+                'name' => 'Name',
+                'note' => 'Note',
+                'order' => 'Order',
             ],
         ],
     ],