+++ csatar-plugins/knowledgerepository/updates/builder_table_create_csatar_knowledgerepository_game_types.php
@@ -13,10 +13,10 @@ class BuilderTableCreateCsatarKnowledgerepositoryGameTypes extends Migration
             $table->increments('id')->unsigned();
             $table->string('name')->nullable();
             $table->string('note')->nullable();
-            $table->integer('order')->nullable()->unsigned();
+            $table->integer('sort_order')->nullable()->unsigned();
         });
     }
-    
+
     public function down()
     {
         Schema::dropIfExists('csatar_knowledgerepository_game_types');